requirements.txt
python-telegram-bot>=20.0
python-dotenv>=1.0.0
requests>=2.31.0
supabase>=2.0.0
pillow>=9.0.0
aiohttp>=3.8.0
pydantic>=2.0.0
loguru>=0.7.0
fastapi>=0.100.0
uvicorn>=0.25.0 